requirements.txt
Flask
numpy
arxiv
scikit-learn
pytz
tqdm
google-generativeai